// src/app/commons/models/types.ts
export type TypeSnackBarPositionHorizontal = 'start' | 'center' | 'end' | 'left' | 'right';
export type TypeSnackBarPositionVertical = 'top' | 'bottom';
export type TypeSnackBarPosition =
	| 'top left'
	| 'top right'
	| 'top center'
	| 'bottom left'
	| 'bottom right'
	| 'bottom center';
export type TypeAnime = 'pokemon' | 'digimon' | 'medabot';